+++ src/classes/EBO.cpp
@@ -8,7 +8,7 @@ EBO::~EBO()
 {
 }
 
-void	EBO::init(GLfloat *indices, GLsizeiptr size)
+void	EBO::init(GLuint *indices, GLsizeiptr size)
 {
 	// Generate a EBO
 	glGenBuffers(1, &ID);
